+ exception handling

+++ src/main/java/com/example/library/repos/BookRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface BookRepository extends CrudRepository<Book, Long> {
 List<Book> findBooksByNameContainsOrAuthorContains(String name,String author);
 
-Book findBookById(Integer index);
+Book findBookById(Long index);
 }
